refactor: login, header 부분에서 불필요한 컴포넌트, 폴더, 파일 정리
