+++ src/apis/auth.ts
@@ -1,34 +1,39 @@
 import { Dispatch, SetStateAction } from "react";
 
+export const logout = async () => {
+  const res = await fetch("/api/auth/login", { method: "DELETE" });
+  console.log("authentication >>", res);
+};
+
 export const updateNickname = async (
   nickname: string,
+  userId: string | undefined,
   setNicknameDupError: Dispatch<SetStateAction<string>>,
   setIsNicknameAllPassed: Dispatch<SetStateAction<boolean>>
 ) => {
   const res = await fetch("/api/auth/mypage/test", {
-    method: "POST",
+    method: "PATCH",
     headers: {
       "Content-Type": "application/json"
     },
-    body: JSON.stringify(nickname)
+    body: JSON.stringify(nickname, userId)
   });
-
-  // console.log("updateNickname >> ", res);
-
-  if (res.status === 200) {
-    setNicknameDupError("사용 가능한 닉네임 입니다");
-    setIsNicknameAllPassed(true);
-  }
 
   if (res.status === 409) {
     setNicknameDupError("동일한 닉네임이 있습니다.");
     setIsNicknameAllPassed(false);
   }
-
   if (res.status === 500) {
     setNicknameDupError("알 수 없는 에러가 발생했습니다.");
     setIsNicknameAllPassed(false);
   }
+  if (res.status === 200) {
+    setNicknameDupError("사용 가능한 닉네임 입니다");
+    setIsNicknameAllPassed(true);
+  }
+
+  const res2 = await fetch("/api/auth/mypage/test", {});
+  // console.log("updateNickname >> ", res);
 };
 
 export const resetPassword = async (email: string) => {
@@ -42,7 +47,7 @@ export const resetPassword = async (email: string) => {
 
   const result = await res.json();
   if (res.ok) {
-    console.log("resetPassword >> ", result);
+    // console.log("resetPassword >> ", result);
   }
 };
 
@@ -57,6 +62,6 @@ export const updatePassword = async (password: string) => {
 
   const result = await res.json();
   if (res.ok) {
-    console.log("resetPassword >> ", result);
+    // console.log("resetPassword >> ", result);
   }
 };
