--- src/apis/auth/auth.ts
@@ -1,4 +1,0 @@
-export const logout = async () => {
-  const res = await fetch("/api/auth/login", { method: "DELETE" });
-  console.log("authentication >>", res);
-};
