+++ src/app/(auth)/login/page.tsx
@@ -1,11 +1,9 @@
 import LoginContainer from "@/app/(auth)/login/_components/LoginContainer";
-import TestLogin from "@/app/(auth)/login/_components/TestLogin";
 
 function Login() {
   return (
     <div className="flex flex-row">
-      {/* <LoginContainer /> */}
-      <TestLogin />
+      <LoginContainer />
     </div>
   );
 }
