+++ src/app/(main)/mypage/edit/_components/UserEditNickname.tsx
@@ -1,8 +1,9 @@
 "use client";
 
-import { updateNickname } from "@/apis/auth/mypage/mypageInfo";
+import { updateNickname } from "@/apis/auth";
 import { CheckNicknameValidity } from "@/app/(main)/mypage/edit/_components/checkValidity";
-import { ChangeEvent, FormEvent, MouseEventHandler, useState } from "react";
+import { useUserContext } from "@/provider/contexts/UserContext";
+import { MouseEvent, useState } from "react";
 
 function UserEditNickname() {
   const [nickname, setNickname] = useState<string>("");
@@ -15,7 +16,10 @@ function UserEditNickname() {
   // 중복확인 통과할 경우: isNicknameAllPassed === true
   // 중복확인 통과 못 할 경우: isNicknameAllPassed === false
 
-  const handleNicknameChange = async (e: FormEvent<HTMLFormElement>) => {
+  const { user } = useUserContext();
+  const userId: string | undefined = user?.userId;
+
+  const handleNicknameChange = async (e: MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     // 유효성 검사
     // CheckNicknameValidity({ nickname, setNicknameError, setIsNicknameValid });
@@ -28,32 +32,31 @@ function UserEditNickname() {
 
     // 3. 중복확인
     if (isNicknameValid) {
-      await updateNickname(nickname, setNicknameDupError, setIsNicknameAllPassed);
+      await updateNickname(nickname, userId || "", setNicknameDupError, setIsNicknameAllPassed);
     }
   };
 
   return (
     <div className="AuthInputDiv">
       <label>닉네임 변경</label>
-      <form className="AuthInputForm" onSubmit={handleNicknameChange}>
-        <input
-          type="text"
-          value={nickname}
-          placeholder="변경할 닉네임을 입력해주세요"
-          className="AuthInputShort"
-          onChange={(e) => {
-            setNickname(e.target.value);
-            CheckNicknameValidity({ nickname: e.target.value, setNicknameError, setIsNicknameValid });
-            if (isNicknameAllPassed) {
-              setNicknameDupError("");
-              setIsNicknameAllPassed(false);
-            }
-          }}
-        />
-        <button className="AuthDupButton" type="submit">
-          중복확인
-        </button>
-      </form>
+      <input
+        type="text"
+        value={nickname}
+        placeholder="변경할 닉네임을 입력해주세요"
+        className="AuthInputShort"
+        onChange={(e) => {
+          setNickname(e.target.value);
+          CheckNicknameValidity({ nickname: e.target.value, setNicknameError, setIsNicknameValid });
+          if (isNicknameAllPassed) {
+            setNicknameDupError("");
+            setIsNicknameAllPassed(false);
+          }
+        }}
+      />
+
+      <button className="AuthDupButton" onClick={handleNicknameChange}>
+        닉네임 변경하기
+      </button>
       {isNicknameValid ? "" : <p className="AuthStateInfo">{nicknameError}</p>}
       {isNicknameAllPassed ? (
         <p className="AuthStateInfoGreen">{nicknameDupError}</p>
